Commit 3:
Initializing more variables are the inputs from the user using
raw_input. Met the input/output, datatype, 3 strings and 3 numbers
requirements.

Used an array to store number inputs; met the array requirement.

Created a dictionary and used a conditional to set the value of a
variable using the dictionary depending on what the choice of the user
was; met the dictionary requirement.