Commit 4:
Added in the last variables.
Created the for loop to get the addition of all the numbers in the
winning_numbers array.
Created conditional to come up with the final money amount to be won in
the lotto.
Created a function to concatenate the adjective with the unusual job to
create a wacky job.