Commit 1: Simple Form Assignment
Initializing my files for the simple form assignment after creating a
new project in the google app launcher. Created the pages.py file, set
up the class and print out method. Used import on the main.py file for
importing the use of the Page class into main.py. Edited the app.yam
file to have the directory for css.